requirements.txt
gitpython
paramiko
scp
